feat: Add drive(), crash() and repair() methods to Car constructor
in this commit, i added the drive() method which takes a parameter called distance
driven and adds it to the odometer property.
The crash() method also resets the canBeDriven property of a car to false and prints a
return statement
the repair() method resets the canBeDriven property of every car to true, making it possible for it to be driven again!

TASK 2 prototype.js